semantic name changes + relocations
