+++ frontend/src/components/Auth/RegistrationSuccess.jsx
